changelog: Extract "git" module
